added tables and hooks. Now trying to connect them
