--- components/pending-table/data/data.tsx
@@ -1,14 +1,0 @@
-export const verifiedStatus = [
-  {
-    value: "VERIFIED",
-    label: "Verified",
-  },
-  {
-    value: "PENDING",
-    label: "Pending",
-  },
-  {
-    value: "REJECTED",
-    label: "Rejected",
-  },
-];
